// src/oem/main/oem-rules/oem-workflow-rules/oem-workflow-rule.enums/quantifiers.enum.ts
import { EligibleTypeEnum } from '../../../../intermediaries/_oem-products-relationships/oem-products-relationships.enums/eligible-type.enum';

export enum QuantifiersEnum {
  A_LINE_ITEM_IN_A_QUOTE = 'A line item in a quote',
  A_SUBMITTED_QUOTE = 'A submitted quote',
  THE_FOLLOWING_PRODUCTS = 'The following products',
  THE_FOLLOWING_TRANSACTION_ATTRIBUTES = 'The following transaction attributes',
  A_GROSS_MARGIN = 'A Gross margin $',
  A_GROSS_MARGIN_PERCENT = 'A Gross margin %',
  A_RELATIVE_VALUE = 'A Relative value / uplift $',
  A_RELATIVE_VALUE_PERCENT = 'A Relative value / uplift %',
  A_QUANTITY_UNITS = 'A Quantity',
  TOTAL_CHANNEL_INCENTIVES = 'Total Channel Incentives',
  A_NET_VALUE = 'A Net value $',
  A_TERM_LENGTH = 'A Term Length',
  INTERNAL_COMMENTS_FIELD = 'Internal comments field',
  EXTERNAL_COMMENTS_FIELD = 'External comments field',
  A_QUOTE_EXPIRATION = 'A Quote Expiration',
  ANY_QUOTE_ATTRIBUTE = 'Any Quote Attribute',
  A_BILLING_PAYMENT_STRUCTURE = 'A Billing / Payment Structure',
  THE_FOLLOWING_SALES_TEAM = 'The following Sales Team',
  THE_FOLLOWING_PARTNERS = 'The following Partners',
  THE_FOLLOWING_USERS = 'The following Users',
  DEAL_TYPE = 'Deal Type',
  A_START_DATE = 'A Start Date',
  AN_END_DATE = 'An End Date',
  LAST_PRICE_PAID = 'Last Price Paid',
  APPROVAL_EMAILS = 'Approval Emails',

  //sub quantifier
  ADDON = 'Addon',
  UPGRADE = 'Upgrade',
  DOWNGRADE = 'Downgrade',
  EXPAND = 'Expand',

  NET_NEW = 'Net New',
  EXPANSION = 'Expansion',
  EXTENSION_RENEWAL = 'Extension/Renewal',

  CHANNEL = 'CHANNEL',
  DIRECT = 'DIRECT',

  RAW_TEXT = 'Raw Text',
  NUMBER = 'Number',
  DOLLARS = 'dollars',
  PERCENT = 'percents',
  UNITS = 'units',
  THE_DEFAULT_SETTINGS = 'The Default Setting',
  DAYS = 'Days',
  YEARS = 'Years',
  DAYS_OUT = 'Days Out',
  DATE = 'Date',
}

//The old one
/*export enum QuantifiersEnum {
  A_LINE_ITEM_IN_A_QUOTE = 'A line item in a quote',
  A_SUBMITTED_QUOTE = 'A submitted quote',
  THE_FOLLOWING_PRODUCTS = 'The following products',
  THE_FOLLOWING_TRANSACTION_ATTRIBUTES = 'The following transaction attributes',
  A_GROSS_MARGIN = 'A gross margin $',
  A_GROSS_MARGIN_PERCENT = 'A gross margin %',
  A_RELATIVE_VALUE = 'A Relative value / uplift $',
  A_RELATIVE_VALUE_PERCENT = 'A Relative value / uplift %',
  A_QUANTITY_UNITS = 'A Quantity',
  A_NET_VALUE = 'A Net value $',
  A_REQUESTED_CUSTOMER_DISCOUNT_PERCENT = 'a requested Customer Discount (%)',
  A_REQUESTED_PARTNER_DISCOUNT_PERCENT = 'a requested Partner Discount (%)',
  A_REQUESTED_DISTRIBUTOR_DISCOUNT_PERCENT = 'a requested Distributor Discount (%)',
  A_TERM_LENGTH = 'a Term Length (days)',
  A_COMMENTS_FIELD = 'a Comments Field',
  A_BILLING_STRUCTURE = 'a Billing Structure',
  THE_FOLLOWING_SALES_TEAM = 'the following Sales Team',
  THE_FOLLOWING_PARTNERS = 'the following Partners',
  THE_FOLLOWING_CUSTOMERS = 'the following Customers',
  DOLLARS = 'dollars',
  PERCENT = 'percents',
  UNITS = 'units',
  THE_DEFAULT_SETTINGS = 'The Default Setting',
  DAYS = 'Days',
  YEARS = 'Years',
  DAYS_OUT = 'Days Out',
  DATE = 'Date',
}*/